backend_api/gateway_service/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
SQLAlchemy==2.0.28
httpx==0.27.0
docker==2.2.1
pydantic-settings==2.2.1
pytest==8.2.2
pyotp
fpdf==1.7.2
PyJWT
pytest-asyncio
loguru
cryptography
passlib
python-jose
redis
web3
py-solc-x
PyYAML
Jinja2
kafka-python
cassandra-driver
pandas==2.3.3
scikit-learn==1.7.2
psycopg2-binary
sqlmodel